requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
langgraph
langchain-aws
boto3
python-dotenv
mcp
dbt-mcp
azure-ai-contentsafety
azure-identity
langchain
